deleted duplicated pages
