--- web/src/components/RedeemPrizeButton 2.tsx
@@ -1,25 +1,0 @@
-const RedeemPrizeButton = () => (
-  <svg
-    width="190"
-    height="53"
-    viewBox="0 0 190 53"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <rect
-      x="0.5"
-      y="0.5"
-      width="189"
-      height="52"
-      rx="9.5"
-      fill="#03045E"
-      stroke="#0676B6"
-    />
-    <path
-      d="M23.7143 30.14C22.9676 30.14 22.2723 30.0187 21.6283 29.776C20.9936 29.524 20.4383 29.174 19.9623 28.726C19.4956 28.2687 19.1316 27.732 18.8703 27.116C18.609 26.5 18.4783 25.828 18.4783 25.1C18.4783 24.372 18.609 23.7 18.8703 23.084C19.1316 22.468 19.5003 21.936 19.9763 21.488C20.4523 21.0307 21.0076 20.6807 21.6423 20.438C22.277 20.186 22.9723 20.06 23.7283 20.06C24.531 20.06 25.2636 20.2 25.9263 20.48C26.589 20.7507 27.149 21.1567 27.6063 21.698L26.4303 22.804C26.0756 22.4213 25.679 22.1367 25.2403 21.95C24.8016 21.754 24.3256 21.656 23.8123 21.656C23.299 21.656 22.8276 21.74 22.3983 21.908C21.9783 22.076 21.6096 22.314 21.2923 22.622C20.9843 22.93 20.7416 23.294 20.5643 23.714C20.3963 24.134 20.3123 24.596 20.3123 25.1C20.3123 25.604 20.3963 26.066 20.5643 26.486C20.7416 26.906 20.9843 27.27 21.2923 27.578C21.6096 27.886 21.9783 28.124 22.3983 28.292C22.8276 28.46 23.299 28.544 23.8123 28.544C24.3256 28.544 24.8016 28.4507 25.2403 28.264C25.679 28.068 26.0756 27.774 26.4303 27.382L27.6063 28.502C27.149 29.034 26.589 29.44 25.9263 29.72C25.2636 30 24.5263 30.14 23.7143 30.14ZM29.1822 30V19.612H30.9322V30H29.1822ZM33.2291 30V22.524H34.9791V30H33.2291ZM34.1111 21.292C33.7844 21.292 33.5137 21.1893 33.2991 20.984C33.0937 20.7787 32.9911 20.5313 32.9911 20.242C32.9911 19.9433 33.0937 19.696 33.2991 19.5C33.5137 19.2947 33.7844 19.192 34.1111 19.192C34.4377 19.192 34.7037 19.29 34.9091 19.486C35.1237 19.6727 35.2311 19.9107 35.2311 20.2C35.2311 20.508 35.1284 20.7693 34.9231 20.984C34.7177 21.1893 34.4471 21.292 34.1111 21.292ZM40.7059 30.098C39.9219 30.098 39.2219 29.9347 38.6059 29.608C37.9993 29.272 37.5233 28.8147 37.1779 28.236C36.8326 27.6573 36.6599 26.9993 36.6599 26.262C36.6599 25.5153 36.8326 24.8573 37.1779 24.288C37.5233 23.7093 37.9993 23.2567 38.6059 22.93C39.2219 22.6033 39.9219 22.44 40.7059 22.44C41.4339 22.44 42.0733 22.5893 42.6239 22.888C43.1839 23.1773 43.6086 23.6067 43.8979 24.176L42.5539 24.96C42.3299 24.6053 42.0546 24.344 41.7279 24.176C41.4106 24.008 41.0653 23.924 40.6919 23.924C40.2626 23.924 39.8753 24.0173 39.5299 24.204C39.1846 24.3907 38.9139 24.6613 38.7179 25.016C38.5219 25.3613 38.4239 25.7767 38.4239 26.262C38.4239 26.7473 38.5219 27.1673 38.7179 27.522C38.9139 27.8673 39.1846 28.1333 39.5299 28.32C39.8753 28.5067 40.2626 28.6 40.6919 28.6C41.0653 28.6 41.4106 28.516 41.7279 28.348C42.0546 28.18 42.3299 27.9187 42.5539 27.564L43.8979 28.348C43.6086 28.908 43.1839 29.342 42.6239 29.65C42.0733 29.9487 41.4339 30.098 40.7059 30.098ZM46.7703 28.334L46.8123 26.136L50.7183 22.524H52.8183L49.5003 25.856L48.5763 26.626L46.7703 28.334ZM45.3423 30V19.612H47.0923V30H45.3423ZM50.9843 30L48.2543 26.626L49.3603 25.226L53.1123 30H50.9843ZM60.9259 30.098C60.1045 30.098 59.4699 29.888 59.0219 29.468C58.5739 29.0387 58.3499 28.4087 58.3499 27.578V20.872H60.0999V27.536C60.0999 27.8907 60.1885 28.166 60.3659 28.362C60.5525 28.558 60.8092 28.656 61.1359 28.656C61.5279 28.656 61.8545 28.5533 62.1159 28.348L62.6059 29.594C62.4005 29.762 62.1485 29.888 61.8499 29.972C61.5512 30.056 61.2432 30.098 60.9259 30.098ZM57.1179 23.98V22.58H62.1019V23.98H57.1179ZM67.1594 30.098C66.394 30.098 65.7127 29.9347 65.1154 29.608C64.518 29.272 64.0467 28.8147 63.7014 28.236C63.356 27.6573 63.1834 26.9993 63.1834 26.262C63.1834 25.5153 63.356 24.8573 63.7014 24.288C64.0467 23.7093 64.518 23.2567 65.1154 22.93C65.7127 22.6033 66.394 22.44 67.1594 22.44C67.934 22.44 68.62 22.6033 69.2174 22.93C69.824 23.2567 70.2954 23.7047 70.6314 24.274C70.9767 24.8433 71.1494 25.506 71.1494 26.262C71.1494 26.9993 70.9767 27.6573 70.6314 28.236C70.2954 28.8147 69.824 29.272 69.2174 29.608C68.62 29.9347 67.934 30.098 67.1594 30.098ZM67.1594 28.6C67.5887 28.6 67.9714 28.5067 68.3074 28.32C68.6434 28.1333 68.9047 27.8627 69.0914 27.508C69.2874 27.1533 69.3854 26.738 69.3854 26.262C69.3854 25.7767 69.2874 25.3613 69.0914 25.016C68.9047 24.6613 68.6434 24.3907 68.3074 24.204C67.9714 24.0173 67.5934 23.924 67.1734 23.924C66.744 23.924 66.3614 24.0173 66.0254 24.204C65.6987 24.3907 65.4374 24.6613 65.2414 25.016C65.0454 25.3613 64.9474 25.7767 64.9474 26.262C64.9474 26.738 65.0454 27.1533 65.2414 27.508C65.4374 27.8627 65.6987 28.1333 66.0254 28.32C66.3614 28.5067 66.7394 28.6 67.1594 28.6ZM76.6919 30V22.524H78.3579V24.582L78.1619 23.98C78.3859 23.476 78.7359 23.0933 79.2119 22.832C79.6973 22.5707 80.2993 22.44 81.0179 22.44V24.106C80.9433 24.0873 80.8733 24.078 80.8079 24.078C80.7426 24.0687 80.6773 24.064 80.6119 24.064C79.9493 24.064 79.4219 24.26 79.0299 24.652C78.6379 25.0347 78.4419 25.6087 78.4419 26.374V30H76.6919ZM85.9481 30.098C85.1175 30.098 84.3895 29.9347 83.7641 29.608C83.1481 29.272 82.6675 28.8147 82.3221 28.236C81.9861 27.6573 81.8181 26.9993 81.8181 26.262C81.8181 25.5153 81.9815 24.8573 82.3081 24.288C82.6441 23.7093 83.1015 23.2567 83.6801 22.93C84.2681 22.6033 84.9355 22.44 85.6821 22.44C86.4101 22.44 87.0588 22.5987 87.6281 22.916C88.1975 23.2333 88.6455 23.6813 88.9721 24.26C89.2988 24.8387 89.4621 25.52 89.4621 26.304C89.4621 26.3787 89.4575 26.4627 89.4481 26.556C89.4481 26.6493 89.4435 26.738 89.4341 26.822H83.2041V25.66H88.5101L87.8241 26.024C87.8335 25.5947 87.7448 25.2167 87.5581 24.89C87.3715 24.5633 87.1148 24.3067 86.7881 24.12C86.4708 23.9333 86.1021 23.84 85.6821 23.84C85.2528 23.84 84.8748 23.9333 84.5481 24.12C84.2308 24.3067 83.9788 24.568 83.7921 24.904C83.6148 25.2307 83.5261 25.618 83.5261 26.066V26.346C83.5261 26.794 83.6288 27.1907 83.8341 27.536C84.0395 27.8813 84.3288 28.1473 84.7021 28.334C85.0755 28.5207 85.5048 28.614 85.9901 28.614C86.4101 28.614 86.7881 28.5487 87.1241 28.418C87.4601 28.2873 87.7588 28.082 88.0201 27.802L88.9581 28.88C88.6221 29.272 88.1975 29.5753 87.6841 29.79C87.1801 29.9953 86.6015 30.098 85.9481 30.098ZM94.3071 30.098C93.5885 30.098 92.9398 29.9393 92.3611 29.622C91.7918 29.2953 91.3438 28.8473 91.0171 28.278C90.6905 27.7087 90.5271 27.0367 90.5271 26.262C90.5271 25.4873 90.6905 24.8153 91.0171 24.246C91.3438 23.6767 91.7918 23.2333 92.3611 22.916C92.9398 22.5987 93.5885 22.44 94.3071 22.44C94.9325 22.44 95.4925 22.58 95.9871 22.86C96.4818 23.1307 96.8738 23.546 97.1631 24.106C97.4525 24.666 97.5971 25.3847 97.5971 26.262C97.5971 27.13 97.4571 27.8487 97.1771 28.418C96.8971 28.978 96.5098 29.398 96.0151 29.678C95.5205 29.958 94.9511 30.098 94.3071 30.098ZM94.5171 28.6C94.9371 28.6 95.3105 28.5067 95.6371 28.32C95.9731 28.1333 96.2391 27.8627 96.4351 27.508C96.6405 27.1533 96.7431 26.738 96.7431 26.262C96.7431 25.7767 96.6405 25.3613 96.4351 25.016C96.2391 24.6613 95.9731 24.3907 95.6371 24.204C95.3105 24.0173 94.9371 23.924 94.5171 23.924C94.0971 23.924 93.7191 24.0173 93.3831 24.204C93.0565 24.3907 92.7905 24.6613 92.5851 25.016C92.3891 25.3613 92.2911 25.7767 92.2911 26.262C92.2911 26.738 92.3891 27.1533 92.5851 27.508C92.7905 27.8627 93.0565 28.1333 93.3831 28.32C93.7191 28.5067 94.0971 28.6 94.5171 28.6ZM96.7851 30V28.236L96.8551 26.248L96.7151 24.26V19.612H98.4511V30H96.7851ZM104.268 30.098C103.438 30.098 102.71 29.9347 102.084 29.608C101.468 29.272 100.988 28.8147 100.642 28.236C100.306 27.6573 100.138 26.9993 100.138 26.262C100.138 25.5153 100.302 24.8573 100.628 24.288C100.964 23.7093 101.422 23.2567 102 22.93C102.588 22.6033 103.256 22.44 104.002 22.44C104.73 22.44 105.379 22.5987 105.948 22.916C106.518 23.2333 106.966 23.6813 107.292 24.26C107.619 24.8387 107.782 25.52 107.782 26.304C107.782 26.3787 107.778 26.4627 107.768 26.556C107.768 26.6493 107.764 26.738 107.754 26.822H101.524V25.66H106.83L106.144 26.024C106.154 25.5947 106.065 25.2167 105.878 24.89C105.692 24.5633 105.435 24.3067 105.108 24.12C104.791 23.9333 104.422 23.84 104.002 23.84C103.573 23.84 103.195 23.9333 102.868 24.12C102.551 24.3067 102.299 24.568 102.112 24.904C101.935 25.2307 101.846 25.618 101.846 26.066V26.346C101.846 26.794 101.949 27.1907 102.154 27.536C102.36 27.8813 102.649 28.1473 103.022 28.334C103.396 28.5207 103.825 28.614 104.31 28.614C104.73 28.614 105.108 28.5487 105.444 28.418C105.78 28.2873 106.079 28.082 106.34 27.802L107.278 28.88C106.942 29.272 106.518 29.5753 106.004 29.79C105.5 29.9953 104.922 30.098 104.268 30.098ZM112.977 30.098C112.147 30.098 111.419 29.9347 110.793 29.608C110.177 29.272 109.697 28.8147 109.351 28.236C109.015 27.6573 108.847 26.9993 108.847 26.262C108.847 25.5153 109.011 24.8573 109.337 24.288C109.673 23.7093 110.131 23.2567 110.709 22.93C111.297 22.6033 111.965 22.44 112.711 22.44C113.439 22.44 114.088 22.5987 114.657 22.916C115.227 23.2333 115.675 23.6813 116.001 24.26C116.328 24.8387 116.491 25.52 116.491 26.304C116.491 26.3787 116.487 26.4627 116.477 26.556C116.477 26.6493 116.473 26.738 116.463 26.822H110.233V25.66H115.539L114.853 26.024C114.863 25.5947 114.774 25.2167 114.587 24.89C114.401 24.5633 114.144 24.3067 113.817 24.12C113.5 23.9333 113.131 23.84 112.711 23.84C112.282 23.84 111.904 23.9333 111.577 24.12C111.26 24.3067 111.008 24.568 110.821 24.904C110.644 25.2307 110.555 25.618 110.555 26.066V26.346C110.555 26.794 110.658 27.1907 110.863 27.536C111.069 27.8813 111.358 28.1473 111.731 28.334C112.105 28.5207 112.534 28.614 113.019 28.614C113.439 28.614 113.817 28.5487 114.153 28.418C114.489 28.2873 114.788 28.082 115.049 27.802L115.987 28.88C115.651 29.272 115.227 29.5753 114.713 29.79C114.209 29.9953 113.631 30.098 112.977 30.098ZM127.608 22.44C128.206 22.44 128.733 22.5567 129.19 22.79C129.657 23.0233 130.021 23.3827 130.282 23.868C130.544 24.344 130.674 24.96 130.674 25.716V30H128.924V25.94C128.924 25.2773 128.78 24.7827 128.49 24.456C128.201 24.1293 127.786 23.966 127.244 23.966C126.862 23.966 126.521 24.05 126.222 24.218C125.933 24.386 125.704 24.6333 125.536 24.96C125.378 25.2867 125.298 25.702 125.298 26.206V30H123.548V25.94C123.548 25.2773 123.404 24.7827 123.114 24.456C122.825 24.1293 122.41 23.966 121.868 23.966C121.486 23.966 121.145 24.05 120.846 24.218C120.557 24.386 120.328 24.6333 120.16 24.96C120.002 25.2867 119.922 25.702 119.922 26.206V30H118.172V22.524H119.838V24.512L119.544 23.924C119.796 23.4387 120.16 23.07 120.636 22.818C121.122 22.566 121.672 22.44 122.288 22.44C122.988 22.44 123.595 22.6127 124.108 22.958C124.631 23.3033 124.976 23.826 125.144 24.526L124.458 24.288C124.701 23.728 125.102 23.28 125.662 22.944C126.222 22.608 126.871 22.44 127.608 22.44ZM140.924 30.098C140.317 30.098 139.762 29.958 139.258 29.678C138.763 29.398 138.367 28.978 138.068 28.418C137.779 27.8487 137.634 27.13 137.634 26.262C137.634 25.3847 137.774 24.666 138.054 24.106C138.343 23.546 138.735 23.1307 139.23 22.86C139.725 22.58 140.289 22.44 140.924 22.44C141.661 22.44 142.31 22.5987 142.87 22.916C143.439 23.2333 143.887 23.6767 144.214 24.246C144.55 24.8153 144.718 25.4873 144.718 26.262C144.718 27.0367 144.55 27.7133 144.214 28.292C143.887 28.8613 143.439 29.3047 142.87 29.622C142.31 29.9393 141.661 30.098 140.924 30.098ZM136.78 32.716V22.524H138.446V24.288L138.39 26.276L138.53 28.264V32.716H136.78ZM140.728 28.6C141.148 28.6 141.521 28.5067 141.848 28.32C142.184 28.1333 142.45 27.8627 142.646 27.508C142.842 27.1533 142.94 26.738 142.94 26.262C142.94 25.7767 142.842 25.3613 142.646 25.016C142.45 24.6613 142.184 24.3907 141.848 24.204C141.521 24.0173 141.148 23.924 140.728 23.924C140.308 23.924 139.93 24.0173 139.594 24.204C139.258 24.3907 138.992 24.6613 138.796 25.016C138.6 25.3613 138.502 25.7767 138.502 26.262C138.502 26.738 138.6 27.1533 138.796 27.508C138.992 27.8627 139.258 28.1333 139.594 28.32C139.93 28.5067 140.308 28.6 140.728 28.6ZM146.377 30V22.524H148.043V24.582L147.847 23.98C148.071 23.476 148.421 23.0933 148.897 22.832C149.383 22.5707 149.985 22.44 150.703 22.44V24.106C150.629 24.0873 150.559 24.078 150.493 24.078C150.428 24.0687 150.363 24.064 150.297 24.064C149.635 24.064 149.107 24.26 148.715 24.652C148.323 25.0347 148.127 25.6087 148.127 26.374V30H146.377ZM152.188 30V22.524H153.938V30H152.188ZM153.07 21.292C152.743 21.292 152.473 21.1893 152.258 20.984C152.053 20.7787 151.95 20.5313 151.95 20.242C151.95 19.9433 152.053 19.696 152.258 19.5C152.473 19.2947 152.743 19.192 153.07 19.192C153.397 19.192 153.663 19.29 153.868 19.486C154.083 19.6727 154.19 19.9107 154.19 20.2C154.19 20.508 154.087 20.7693 153.882 20.984C153.677 21.1893 153.406 21.292 153.07 21.292ZM155.619 30V28.894L160.225 23.322L160.547 23.924H155.703V22.524H161.947V23.63L157.341 29.202L156.991 28.6H162.073V30H155.619ZM167.091 30.098C166.26 30.098 165.532 29.9347 164.907 29.608C164.291 29.272 163.81 28.8147 163.465 28.236C163.129 27.6573 162.961 26.9993 162.961 26.262C162.961 25.5153 163.124 24.8573 163.451 24.288C163.787 23.7093 164.244 23.2567 164.823 22.93C165.411 22.6033 166.078 22.44 166.825 22.44C167.553 22.44 168.201 22.5987 168.771 22.916C169.34 23.2333 169.788 23.6813 170.115 24.26C170.441 24.8387 170.605 25.52 170.605 26.304C170.605 26.3787 170.6 26.4627 170.591 26.556C170.591 26.6493 170.586 26.738 170.577 26.822H164.347V25.66H169.653L168.967 26.024C168.976 25.5947 168.887 25.2167 168.701 24.89C168.514 24.5633 168.257 24.3067 167.931 24.12C167.613 23.9333 167.245 23.84 166.825 23.84C166.395 23.84 166.017 23.9333 165.691 24.12C165.373 24.3067 165.121 24.568 164.935 24.904C164.757 25.2307 164.669 25.618 164.669 26.066V26.346C164.669 26.794 164.771 27.1907 164.977 27.536C165.182 27.8813 165.471 28.1473 165.845 28.334C166.218 28.5207 166.647 28.614 167.133 28.614C167.553 28.614 167.931 28.5487 168.267 28.418C168.603 28.2873 168.901 28.082 169.163 27.802L170.101 28.88C169.765 29.272 169.34 29.5753 168.827 29.79C168.323 29.9953 167.744 30.098 167.091 30.098Z"
-      fill="#EFF8FA"
-    />
-  </svg>
-);
-
-export default RedeemPrizeButton;
