--- web/src/pages/Landing-Page 2.tsx
@@ -1,48 +1,0 @@
-import WDCCLogo from "../assets/WDCC_Logo.svg";
-import { useNavigate } from "react-router";
-import styles from "../styles/page styles/Landing-Page.module.css";
-
-export const HomePage = () => {
-  const navigate = useNavigate();
-  const handleButtonClick = () => {
-    navigate("/sign-in");
-  };
-  return (
-    <div className={styles.background}>
-      <title>WDCC Passport</title>
-      <img
-        className="absolute top-[0.1em] md:top-[3em] right-[1.5em] md:right-[3em] w-[9em] h-auto"
-        src={WDCCLogo}
-        alt="WDCC Logo"
-      />
-      <div className="flex flex-col-reverse md:flex-row gap-20 max-w-7xl mx-auto items-center justify-center p-12 lg:p-24 rounded-xl">
-        <div className="flex flex-col gap-16">
-          <h1 className="text-[#e1ebff] text-4xl md:text-5xl lg:text-6xl font-display font-bold">
-            WDCC passport
-          </h1>
-
-          <div className="text-[#e1ebff] text-lg font-semibold opacity-90 font-sans flex flex-col gap-6">
-            <p>
-              {" "}
-              Unlock a world of engagement with WDCC Passport!
-            </p>
-            <p>
-              {" "}
-              Scan, attend, and collect rewards as you embark on your journey through WDCC events.
-            </p>
-            <a href="/team">Developed by WDCC. Meet our Team!</a>
-            <a className="text-sm" href="/privacy-policy">
-              Our Privacy Policy
-            </a>
-          </div>
-          <div>
-            <button
-              onClick={handleButtonClick}
-              className="bg-[#e1ebff] text-[#03045e] font-semibold text-lg py-2 px-4 rounded-lg">
-                Get Started </button>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-};
