--- web/src/styles/page styles/Landing-Page.module 2.css
@@ -1,9 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Open+Sans:ital,wght@0,300;0,400;1,300;1,400&display=swap');
-
-.background {
-    min-height: 100vh;
-    background-color: #03045e;
-    align-items: center;
-    justify-content: center;
-    display: flex;
-  }
