--- web/src/styles/page styles/Leaderboard 2.css
@@ -1,4 +1,0 @@
-.leaderboard-main {
-    background-color: #e1ebff;
-    color: #03045e;
-}
