+++ web/src/styles/page styles/Leaderboard.css
@@ -1,4 +1,4 @@
-.passport-main {
+.leaderboard-main {
     background-color: #e1ebff;
     color: #03045e;
 }
